Fix slot bug for pets
This fix the pets that dissapear when you change your dual spec,
dismount or some things that causes that you lose your pet. This is a
roc13x fix improved by me, because the roc13x fix causes crashes on
Battlegrounds when the hunter is resurrected by a Spirit Healer. Tested
and works perfectly. Enjoy it :D